+++ backend/src/test/java/com/woowacourse/zzimkkong/controller/AdminControllerTest.java
@@ -11,7 +11,7 @@ import com.woowacourse.zzimkkong.dto.reservation.ReservationResponse;
 import com.woowacourse.zzimkkong.dto.space.SpaceFindDetailWithIdResponse;
 import com.woowacourse.zzimkkong.infrastructure.auth.AuthorizationExtractor;
 import com.woowacourse.zzimkkong.service.AdminService;
-import com.woowacourse.zzimkkong.service.slack.SlackService;
+import com.woowacourse.zzimkkong.service.SlackService;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
