+++ backend/src/test/java/com/woowacourse/zzimkkong/controller/ManagerReservationControllerTest.java
@@ -3,7 +3,7 @@ package com.woowacourse.zzimkkong.controller;
 import com.woowacourse.zzimkkong.domain.*;
 import com.woowacourse.zzimkkong.dto.reservation.*;
 import com.woowacourse.zzimkkong.infrastructure.auth.AuthorizationExtractor;
-import com.woowacourse.zzimkkong.service.slack.SlackService;
+import com.woowacourse.zzimkkong.service.SlackService;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
